start on mod interface rewrite
