--- templates/login.html
@@ -1,27 +1,0 @@
-{% if error %}<h2 style="text-align:center">{{ error }}</h2>{% endif %}
-<form action="" method="post">
-{% if redirect %}<input type="hidden" name="redirect" value="{{ redirect }}" />{% endif %}
-<table style="margin-top:25px;">
-	<tr>
-		<th>
-			{% trans %}Username{% endtrans %}
-		</th>
-		<td>
-			<input type="text" name="username" size="20" maxlength="30" value="{{ username }}">
-		</td>
-	</tr>
-	<tr>
-		<th>
-			{% trans %}Password{% endtrans %}
-		</th>
-		<td>
-			<input type="password" name="password" size="20" maxlength="30" value="">
-		</td>
-	</tr>
-	<tr>
-		<td></td>
-		<td>
-			<input type="submit" name="login" value="{% trans %}Continue{% endtrans %}" />
-		</td>
-</table>
-</form>
